Moved the graphics to separate module
